+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/__init__.py
@@ -44,6 +44,7 @@ from ._models import (
     CurrencyValue,
     CustomDocumentModelsDetails,
     ModelBuildMode,
+    DocumentClassifierDetails,
     DocumentField,
     DocumentKeyValuePair,
     DocumentKeyValueElement,
@@ -67,6 +68,11 @@ from ._models import (
     DocumentAnalysisInnerError,
     TargetAuthorization,
     QuotaDetails,
+)
+from ._generated.models import (  # patched models
+    ClassifierDocumentTypeDetails,
+    AzureBlobFileListSource,
+    AzureBlobContentSource,
 )
 from ._api_versions import FormRecognizerApiVersion, DocumentAnalysisApiVersion
 
@@ -111,6 +117,7 @@ __all__ = [
     "CurrencyValue",
     "CustomDocumentModelsDetails",
     "ModelBuildMode",
+    "DocumentClassifierDetails",
     "DocumentField",
     "DocumentKeyValueElement",
     "DocumentKeyValuePair",
@@ -135,6 +142,9 @@ __all__ = [
     "DocumentAnalysisError",
     "DocumentAnalysisInnerError",
     "TargetAuthorization",
+    "ClassifierDocumentTypeDetails",
+    "AzureBlobFileListSource",
+    "AzureBlobContentSource",
     "QuotaDetails",
 ]