+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -6,6 +6,8 @@
 - Added `features` keyword argument on `begin_analyze_document()` and `begin_analyze_document_from_url()`.
 - Added `AnalysisFeature` enum with optional document analysis feature to enable.
 - Added the following optional properties on `DocumentStyle` class: `similar_font_family`, `font_style`, `font_weight`, `color`, `background_color`.
+- Added support for custom document classification on `DocumentModelAdministrationClient`: `begin_build_document_classifier`, 
+`list_document_classifiers`, `get_document_classifier`, and `delete_document_classifier`.
 - Added model `QuotaDetails` and property `custom_neural_document_model_builds` on `ResourceDetails`.
 - Added kind `documentClassifierBuild` to `OperationSummary`.
 - Added property `expires_on` to `DocumentModelDetails` and `DocumentModelSummary`.