+++ sdk/storage/azure-storage-file-share/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.12.0b1 (Unreleased)
+## 12.12.0b1 (2023-03-28)
 
 ### Features Added
 - Added support for service version 2022-11-02.